% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mCB.r
\name{cvCB}
\alias{cvCB}
\title{Coefficient of Variation of the Contracted-Beta Distribution}
\usage{
cvCB(alpha, beta)
}
\arguments{
\item{alpha}{Positive shape parameter.}

\item{beta}{Positive shape parameter.}
}
\value{
A numeric value.
}
\description{
Computes the theoretical coefficient of variation of the CB distribution.
}
\examples{
cvCB(alpha = 2, beta = 5)

}
