% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mCB.r
\name{kurtCB}
\alias{kurtCB}
\title{Kurtosis of the Contracted-Beta Distribution}
\usage{
kurtCB(alpha, beta, excess = TRUE)
}
\arguments{
\item{alpha}{Positive shape parameter.}

\item{beta}{Positive shape parameter.}

\item{excess}{Logical. If \code{TRUE}, the excess kurtosis is returned;
otherwise, the ordinary kurtosis is returned.}
}
\value{
A numeric value.
}
\description{
Computes the theoretical kurtosis or excess kurtosis of the CB distribution.
}
\examples{
kurtCB(alpha = 2, beta = 5)
kurtCB(alpha = 2, beta = 5, excess = FALSE)

}
