% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mCB.r
\name{mCB}
\alias{mCB}
\title{Raw Moments of the Contracted-Beta Distribution}
\usage{
mCB(r, alpha, beta)
}
\arguments{
\item{r}{Non-negative order or vector of orders of the moment.}

\item{alpha}{Positive shape parameter.}

\item{beta}{Positive shape parameter.}
}
\value{
A numeric vector with the raw moments of order \eqn{r}.
}
\description{
Computes the raw moment of order \eqn{r} for the
Contracted-Beta (CB) distribution.
}
\details{
The r-th raw moment is given by

\deqn{
\mathbb{E}(T^r)
=
\frac{B(\alpha + r, \beta)}{B(\alpha, \beta)}
\times
c_r,
}

where

\deqn{
c_r =
\left\{
\begin{array}{ll}
\log(2), & r = 1, \\
\dfrac{1 - 2^{1-r}}{r - 1}, & r \neq 1.
\end{array}
\right.
}
}
\examples{
mCB(1, alpha = 2, beta = 5)
mCB(1:4, alpha = 2, beta = 5)

}
