% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mCB.r
\name{meanCB}
\alias{meanCB}
\title{Mean of the Contracted-Beta Distribution}
\usage{
meanCB(alpha, beta)
}
\arguments{
\item{alpha}{Positive shape parameter.}

\item{beta}{Positive shape parameter.}
}
\value{
A numeric value.
}
\description{
Computes the theoretical mean of the CB distribution.
}
\examples{
meanCB(alpha = 2, beta = 5)

}
