% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mCB.r
\name{skewCB}
\alias{skewCB}
\title{Skewness of the Contracted-Beta Distribution}
\usage{
skewCB(alpha, beta)
}
\arguments{
\item{alpha}{Positive shape parameter.}

\item{beta}{Positive shape parameter.}
}
\value{
A numeric value.
}
\description{
Computes the theoretical skewness coefficient of the CB distribution.
}
\examples{
skewCB(alpha = 2, beta = 5)

}
