% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mCB.r
\name{varCB}
\alias{varCB}
\title{Variance of the Contracted-Beta Distribution}
\usage{
varCB(alpha, beta)
}
\arguments{
\item{alpha}{Positive shape parameter.}

\item{beta}{Positive shape parameter.}
}
\value{
A numeric value.
}
\description{
Computes the theoretical variance of the CB distribution.
}
\examples{
varCB(alpha = 2, beta = 5)

}
